+++ Src/mainmenu.c
@@ -14,7 +14,6 @@
 #include "utils.h"
 
 #include "mainmenu.h"
-
 #include "play.h"
 #include "howToPlay1.h"
 #include "credits1.h"
@@ -51,7 +50,7 @@ CP_Image hoverStartWord = NULL;
 void mainMenu_Init()
 {   
     /*------------------------------------------------------------ Window Size 1000 x 800 ------------------------------------------------------------*/
-    CP_System_SetWindowSize(windowWidth, windowHeight);
+    CP_System_SetWindowSize((int)windowWidth, (int)windowHeight);
 
     CP_System_SetWindowTitle("CANT CEE");
 
@@ -102,73 +101,73 @@ void mainMenu_Update()
     CP_Image_Draw(bracket, windowWidth / 2, windowHeight / 2, windowWidth, windowHeight, 255);
     CP_Image_Draw(gradientAccentBackground, windowWidth / 2, windowHeight / 2, windowWidth, windowHeight, 255);
 
-    CP_Image_Draw(startWord, startWordWidth, startWordHeight, CP_Image_GetWidth(startWord), CP_Image_GetHeight(startWord), 255);
-    CP_Image_Draw(howToPlayWord, howToPlayWordWidth, howToPlayWordHeight, CP_Image_GetWidth(howToPlayWord), CP_Image_GetHeight(howToPlayWord), 255);
-    CP_Image_Draw(creditsWord, creditsWordWidth, creditsWordHeight, CP_Image_GetWidth(creditsWord), CP_Image_GetHeight(creditsWord), 255);
-    CP_Image_Draw(settingsWord, settingsWordWidth, settingsWordHeight, CP_Image_GetWidth(settingsWord), CP_Image_GetHeight(settingsWord), 255);
-    CP_Image_Draw(exitWord, exitWordWidth, exitWordHeight, CP_Image_GetWidth(exitWord), CP_Image_GetHeight(exitWord), 255);
+    CP_Image_Draw(startWord, startWordWidth, startWordHeight, (float)CP_Image_GetWidth(startWord), (float)CP_Image_GetHeight(startWord), 255);
+    CP_Image_Draw(howToPlayWord, howToPlayWordWidth, howToPlayWordHeight, (float)CP_Image_GetWidth(howToPlayWord), (float)CP_Image_GetHeight(howToPlayWord), 255);
+    CP_Image_Draw(creditsWord, creditsWordWidth, creditsWordHeight, (float)CP_Image_GetWidth(creditsWord), (float)CP_Image_GetHeight(creditsWord), 255);
+    CP_Image_Draw(settingsWord, settingsWordWidth, settingsWordHeight, (float)CP_Image_GetWidth(settingsWord), (float)CP_Image_GetHeight(settingsWord), 255);
+    CP_Image_Draw(exitWord, exitWordWidth, exitWordHeight, (float)CP_Image_GetWidth(exitWord), (float)CP_Image_GetHeight(exitWord), 255);
 
     // ---------------------------------------------------------------Mouse Inputs/Checking---------------------------------------------------------------
     // --------------- Hover Feature ---------------
-    if (IsAreaClicked(startWordWidth, startWordHeight, CP_Image_GetWidth(startWord), CP_Image_GetHeight(startWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+    if (IsAreaClicked(startWordWidth, startWordHeight, (float)CP_Image_GetWidth(startWord), (float)CP_Image_GetHeight(startWord), (float)CP_Input_GetMouseX(), (float)CP_Input_GetMouseY()))
     {
-        CP_Image_Draw(hoverStart, startWordWidth - 5, startWordHeight - 5, CP_Image_GetWidth(startWord), CP_Image_GetHeight(startWord), 255);
-        CP_Image_Draw(hoverStartWord, startWordWidth, startWordHeight, CP_Image_GetWidth(startWord), CP_Image_GetHeight(startWord), 255);
+        CP_Image_Draw(hoverStart, startWordWidth - 5, startWordHeight - 5, (float)CP_Image_GetWidth(startWord), (float)CP_Image_GetHeight(startWord), 255);
+        CP_Image_Draw(hoverStartWord, startWordWidth, startWordHeight, (float)CP_Image_GetWidth(startWord), (float)CP_Image_GetHeight(startWord), 255);
     }
 
-    if (IsAreaClicked(howToPlayWordWidth, howToPlayWordHeight, CP_Image_GetWidth(hoverHowToPLay), CP_Image_GetHeight(hoverHowToPLay), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+    if (IsAreaClicked(howToPlayWordWidth, howToPlayWordHeight, (float)CP_Image_GetWidth(hoverHowToPLay), (float)CP_Image_GetHeight(hoverHowToPLay), (float)CP_Input_GetMouseX(), (float)CP_Input_GetMouseY()))
     {
-        CP_Image_Draw(hoverHowToPLay, howToPlayWordWidth - 5, howToPlayWordHeight - 5, CP_Image_GetWidth(hoverHowToPLay), CP_Image_GetHeight(hoverHowToPLay), 255);
+        CP_Image_Draw(hoverHowToPLay, howToPlayWordWidth - 5, howToPlayWordHeight - 5, (float)CP_Image_GetWidth(hoverHowToPLay), (float)CP_Image_GetHeight(hoverHowToPLay), 255);
     }
 
-    if (IsAreaClicked(creditsWordWidth, creditsWordHeight, CP_Image_GetWidth(creditsWord), CP_Image_GetHeight(creditsWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+    if (IsAreaClicked(creditsWordWidth, creditsWordHeight, (float)CP_Image_GetWidth(creditsWord), (float)CP_Image_GetHeight(creditsWord), (float)CP_Input_GetMouseX(), (float)CP_Input_GetMouseY()))
     {
-        CP_Image_Draw(hoverCredits, creditsWordWidth - 5, creditsWordHeight - 5, CP_Image_GetWidth(hoverCredits), CP_Image_GetHeight(hoverCredits), 255);
+        CP_Image_Draw(hoverCredits, creditsWordWidth - 5, creditsWordHeight - 5, (float)CP_Image_GetWidth(hoverCredits), (float)CP_Image_GetHeight(hoverCredits), 255);
     }
 
-    if (IsAreaClicked(settingsWordWidth, settingsWordHeight, CP_Image_GetWidth(creditsWord), CP_Image_GetHeight(creditsWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+    if (IsAreaClicked(settingsWordWidth, settingsWordHeight, (float)CP_Image_GetWidth(creditsWord), (float)CP_Image_GetHeight(creditsWord), (float)CP_Input_GetMouseX(), (float)CP_Input_GetMouseY()))
     {
-        CP_Image_Draw(hoverSettings, settingsWordWidth - 5, settingsWordHeight - 5, CP_Image_GetWidth(hoverSettings), CP_Image_GetHeight(hoverSettings), 255);
+        CP_Image_Draw(hoverSettings, settingsWordWidth - 5, settingsWordHeight - 5, (float)CP_Image_GetWidth(hoverSettings), (float)CP_Image_GetHeight(hoverSettings), 255);
     }
 
-    if (IsAreaClicked(exitWordWidth, exitWordHeight, CP_Image_GetWidth(creditsWord), CP_Image_GetHeight(creditsWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+    if (IsAreaClicked(exitWordWidth, exitWordHeight, (float)CP_Image_GetWidth(creditsWord), (float)CP_Image_GetHeight(creditsWord), (float)CP_Input_GetMouseX(), (float)CP_Input_GetMouseY()))
     {
-        CP_Image_Draw(hoverExit, exitWordWidth - 5, exitWordHeight - 5, CP_Image_GetWidth(hoverExit), CP_Image_GetHeight(hoverExit), 255);
+        CP_Image_Draw(hoverExit, exitWordWidth - 5, exitWordHeight - 5, (float)CP_Image_GetWidth(hoverExit), (float)CP_Image_GetHeight(hoverExit), 255);
     }
 
     // --------------- Mouse Input ---------------
     if (CP_Input_MouseTriggered(MOUSE_BUTTON_LEFT) == 1)
     {
         // ----------Play Button Input----------
-        if (IsAreaClicked(startWordWidth, startWordHeight, CP_Image_GetWidth(startWord), CP_Image_GetHeight(startWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+        if (IsAreaClicked(startWordWidth, startWordHeight, (float)CP_Image_GetWidth(startWord), (float)CP_Image_GetHeight(startWord), (float)CP_Input_GetMouseX(), (float)CP_Input_GetMouseY()))
         {
             CP_Sound_PlayAdvanced(buttonSound, CP_Sound_GetGroupVolume(CP_SOUND_GROUP_SFX), 2.0f, FALSE, CP_SOUND_GROUP_SFX);
             CP_Engine_SetNextGameStateForced(play_Init, play_Update, play_Exit); // play.c > goes to level select > click level to play
         }
 
         // ----------How to Play Input----------
-        if (IsAreaClicked(howToPlayWordWidth, howToPlayWordHeight, CP_Image_GetWidth(howToPlayWord), CP_Image_GetHeight(howToPlayWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+        if (IsAreaClicked(howToPlayWordWidth, howToPlayWordHeight, (float)CP_Image_GetWidth(howToPlayWord), (float)CP_Image_GetHeight(howToPlayWord), (float)CP_Input_GetMouseX(), (float)CP_Input_GetMouseY()))
         {
             CP_Sound_PlayAdvanced(buttonSound, CP_Sound_GetGroupVolume(CP_SOUND_GROUP_SFX), 2.0f, FALSE, CP_SOUND_GROUP_SFX);
             CP_Engine_SetNextGameStateForced(howToPlay1_Init, howToPlay1_Update, howToPlay1_Exit);
         }
 
         // ----------Credit Input----------
-        if (IsAreaClicked(creditsWordWidth, creditsWordHeight, CP_Image_GetWidth(creditsWord), CP_Image_GetHeight(creditsWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+        if (IsAreaClicked(creditsWordWidth, creditsWordHeight, (float)CP_Image_GetWidth(creditsWord), (float)CP_Image_GetHeight(creditsWord), (float)CP_Input_GetMouseX(), (float)CP_Input_GetMouseY()))
         {
             CP_Sound_PlayAdvanced(buttonSound, CP_Sound_GetGroupVolume(CP_SOUND_GROUP_SFX), 2.0f, FALSE, CP_SOUND_GROUP_SFX);
             CP_Engine_SetNextGameStateForced(credits1_Init, credits1_Update, credits1_Exit);
         }
 
         // ----------Settings Input----------
-        if (IsAreaClicked(settingsWordWidth, settingsWordHeight, CP_Image_GetWidth(settingsWord), CP_Image_GetHeight(settingsWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+        if (IsAreaClicked(settingsWordWidth, settingsWordHeight, (float)CP_Image_GetWidth(settingsWord), (float)CP_Image_GetHeight(settingsWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
         {
             CP_Sound_PlayAdvanced(buttonSound, CP_Sound_GetGroupVolume(CP_SOUND_GROUP_SFX), 2.0f, FALSE, CP_SOUND_GROUP_SFX);
             CP_Engine_SetNextGameStateForced(settings_Init, settings_Update, settings_Exit);
         }
 
         // ----------Exit Button Input----------
-        if (IsAreaClicked(exitWordWidth, exitWordHeight, CP_Image_GetWidth(exitWord), CP_Image_GetHeight(exitWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
+        if (IsAreaClicked(exitWordWidth, exitWordHeight, (float)CP_Image_GetWidth(exitWord), (float)CP_Image_GetHeight(exitWord), CP_Input_GetMouseX(), CP_Input_GetMouseY()))
         {
             CP_Sound_PlayAdvanced(buttonSound, CP_Sound_GetGroupVolume(CP_SOUND_GROUP_SFX), 2.0f, FALSE, CP_SOUND_GROUP_SFX);
             CP_Engine_Terminate();
